pyproject.toml
[tool.poetry]
name = "ml-system-design"
version = "0.1.0"
description = "End to End ML Deployment"
authors = ["Aditya Rawat"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
langchain-core = "^0.2.7"
langchain-openai = "^0.1.8"
streamlit = "^1.35.0"
python-dotenv = "^1.0.1"
tenacity = "8.3.0"
watchdog = "^4.0.1"
langserve = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pre-commit = "^2.20.0"
mypy = "^0.991"
flake8 = "^6.0.0"
black = "^22.10.0"
freezegun = "^1.2.2"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
